Moved all connector functions calling income-tax-view-change into a single connector, updated tests and fixed issues
