+++ app/config/DIModule.scala
@@ -18,10 +18,12 @@ package config
 
 import auth.FrontendAuthorisedFunctions
 import com.google.inject.AbstractModule
+import connectors.{IncomeTaxViewChangeConnector, IncomeTaxViewChangeConnectorImpl}
 import uk.gov.hmrc.auth.core.AuthorisedFunctions
 
 class DIModule extends AbstractModule{
   def configure(): Unit = {
     bind(classOf[AuthorisedFunctions]).to(classOf[FrontendAuthorisedFunctions]).asEagerSingleton()
+    bind(classOf[IncomeTaxViewChangeConnector]).to(classOf[IncomeTaxViewChangeConnectorImpl]).asEagerSingleton()
   }
 }
