--- app/connectors/ReportDeadlinesConnector.scala
@@ -1,82 +1,0 @@
-/*
- * Copyright 2019 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package connectors
-
-import javax.inject.{Inject, Singleton}
-import audit.AuditingService
-import audit.models.{ReportDeadlinesRequestAuditModel, ReportDeadlinesResponseAuditModel}
-import auth.{MtdItUser, MtdItUserWithNino}
-import config.FrontendAppConfig
-import models.reportDeadlines.{ReportDeadlinesErrorModel, ReportDeadlinesModel, ReportDeadlinesResponseModel}
-import play.api.Logger
-import play.api.http.Status
-import play.api.http.Status.OK
-import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
-import uk.gov.hmrc.play.bootstrap.http.HttpClient
-import uk.gov.hmrc.play.http.logging.MdcLoggingExecutionContext._
-
-import scala.concurrent.Future
-
-@Singleton
-class ReportDeadlinesConnector @Inject()(val http: HttpClient,
-                                         val config: FrontendAppConfig,
-                                         val auditingService: AuditingService
-                                        ) extends RawResponseReads {
-
-  private[connectors] def getReportDeadlinesUrl(incomeSourceID: String, nino: String): String =
-    s"${config.itvcProtectedService}/income-tax-view-change/$nino/income-source/$incomeSourceID/report-deadlines"
-
-  def getReportDeadlines(incomeSourceID: String)(implicit headerCarrier: HeaderCarrier, mtdUser: MtdItUser[_]): Future[ReportDeadlinesResponseModel] = {
-
-    val url = getReportDeadlinesUrl(incomeSourceID, mtdUser.nino)
-    Logger.debug(s"[ReportDeadlinesConnector][getReportDeadlines] - GET $url")
-
-    //Audit Report Deadlines Request
-    auditingService.audit(ReportDeadlinesRequestAuditModel(mtdUser.mtditid, mtdUser.nino, incomeSourceID))
-
-    http.GET[HttpResponse](url)(httpReads, headerCarrier, implicitly) map {
-      response =>
-        response.status match {
-          case OK =>
-            Logger.debug(s"[ReportDeadlinesConnector][getReportDeadlines] - RESPONSE status: ${response.status}, json: ${response.json}")
-            response.json.validate[ReportDeadlinesModel].fold(
-              invalid => {
-                Logger.error("[ReportDeadlinesConnector][getReportDeadlines] - Json Validation Error. Parsing Report Deadlines Data Response")
-                Logger.error(s"[ReportDeadlinesConnector][getReportDeadlines] - Json Validation Error: $invalid")
-                ReportDeadlinesErrorModel(Status.INTERNAL_SERVER_ERROR, "Json Validation Error. Parsing Report Deadlines Data Response")
-              },
-              valid => {
-                //Audit Report Deadlines Response
-                auditingService.extendedAudit(ReportDeadlinesResponseAuditModel(mtdUser.mtditid, mtdUser.nino, incomeSourceID, valid.obligations))
-                valid
-              }
-            )
-          case _ =>
-            Logger.error(s"[ReportDeadlinesConnector][getReportDeadlines] - RESPONSE status: ${response.status}, body: ${response.body}")
-            Logger.error(s"[ReportDeadlinesConnector][getReportDeadlines] - Status: [${response.status}] Returned from business report deadlines call")
-            ReportDeadlinesErrorModel(response.status, response.body)
-        }
-    } recover {
-      case ex =>
-        Logger.error(s"[ReportDeadlinesConnector][getReportDeadlines] - Unexpected future failed error, ${ex.getMessage}")
-        ReportDeadlinesErrorModel(Status.INTERNAL_SERVER_ERROR, s"Unexpected future failed error, ${ex.getMessage}")
-      case _ =>
-        Logger.error("[ReportDeadlinesConnector][getReportDeadlines] - Unexpected future failed error")
-        ReportDeadlinesErrorModel(Status.INTERNAL_SERVER_ERROR, "Unexpected future failed error")
-    }
-  }
-}
