+++ test/services/CalculationServiceSpec.scala
@@ -19,19 +19,16 @@ package services
 import assets.BaseTestConstants._
 import assets.CalcBreakdownTestConstants._
 import assets.EstimatesTestConstants._
-import mocks.connectors.{MockCalculationDataConnector, MockLastTaxCalculationConnector}
+import mocks.connectors.{MockCalculationDataConnector, MockIncomeTaxViewChangeConnector}
 import mocks.services.MockCalculationService
 import models.calculation._
 import play.api.http.Status
 import testUtils.TestSupport
-import uk.gov.hmrc.http.HeaderCarrier
 
-import scala.concurrent.Future
-
-class CalculationServiceSpec extends TestSupport with MockLastTaxCalculationConnector with MockCalculationDataConnector with MockCalculationService{
+class CalculationServiceSpec extends TestSupport with MockIncomeTaxViewChangeConnector with MockCalculationDataConnector with MockCalculationService{
 
   object TestCalculationService extends CalculationService(
-    mockLastTaxCalculationConnector,
+    mockIncomeTaxViewChangeConnector,
     mockCalculationDataConnector,
     frontendAppConfig
   )
@@ -78,7 +75,7 @@ class CalculationServiceSpec extends TestSupport with MockLastTaxCalculationConn
   "The CalculationService.getAllLatestCalculations method" when {
 
     object TestCalculationService extends CalculationService(
-      mockLastTaxCalculationConnector,
+      mockIncomeTaxViewChangeConnector,
       mockCalculationDataConnector,
       frontendAppConfig
     )
